backend/requirements.txt
Flask==2.0.1
gunicorn==20.1.0
sqlalchemy==1.4.23
pymysql==1.0.2
requests==2.26.0
chess==1.6.1
tensorflow==2.6.0
pyjwt==2.1.0